Fixed webapp sometimes not disabling when closing the client Fair amount of cleanup in WSClient

+++ src/main/java/com/jcs/overlay/websocket/messages/J2W/SetSummonerSpellsMessage.java
@@ -3,10 +3,10 @@ package com.jcs.overlay.websocket.messages.J2W;
 public class SetSummonerSpellsMessage {
     private final String messageType = "SetSummonerSpells";
     private final long cellId;
-    private final byte spellSlot;
+    private final int spellSlot;
     private final Long spellId;
 
-    public SetSummonerSpellsMessage(long adjustedCellId, byte spellSlot, Long spellId) {
+    public SetSummonerSpellsMessage(long adjustedCellId, int spellSlot, Long spellId) {
         this.cellId = adjustedCellId;
         this.spellSlot = spellSlot;
         this.spellId = spellId;
